Create Matrix processor
This is a study task: to create the calculator that will do easy operations with two-dimensional matrices: addition, multiplication to a scalar, multiplication of two matrices, transposition. A funny part about transposition - students were tasked to perform also side diagonal transposition, vertical symmetry, and horizontal symmetry transposition: in fact, just to figure out better, how the "range" function works.
I want that the input reader can guess whether the string contains a decimal dot or not, and convert it according to that to integer (if no decimal point) or to float (if it's there). For now, I know Try - Except method, which, according to forums, is slow (to check). Otherwise, I tried list comprehension with if - else conditioning, but it doesn't work for the moment.