Create Matrix processor.py
This project makes easy operations with two-dimensional matrices (something like a very simplified version of NumPy). 
A funny feature of transposition of a matrix - students were tasked to do also transposition to side diagonal, to horizontal and vertical lines.
The Try-Except method is a good way to read an input string and decide whether to convert it to integer or to a string. But here in my project, I tried to use list comprehension with conditions if-else. For the moment there is a bug, I will discover later how to make it work.