Create Huffman_encoding_algorithme
Huffman encoding algorithme turns a string into a binary string. The code corresponding to every symbol has a variable length, depending on how frequent the symbol is in a given string. My algorithm creates a binary tree. Each of its branches leads to zero or one, that creates a full code while passing from the root to the leaves.
This task was given to me during the free course "Algorithms. Theory, practice, methods" on the self-educational platform stepik.org