Create Huffman decoding, no tree
Reconstruct a string from its non-prefix binaty code and letter-code listing.
Given number of various letters in the encoded string k, and the length of the encoded string. Next k lines are given the "letter: binary code". Then encoded string. 
I made this algorithm just with search in a dictionary, not creating the tree. The solution is still fast enough to pass the test with the given time limit (size of the string is within 10 000 symbols).